[sclang] Move QcCallback declaration
QcCallback was defined in the QcWebView header. Since QcWebView will be made optional, this commit moves QcCallback out.

Also removes a dangerous "using" in header file.
